refactor(cli): split yaml_pattern_registry out of yaml_generators
YamlPattern, PatternScale, and ALL_PATTERNS move to their own file so
they can be shared without pulling in every generate_* function (and
rand/rand_chacha). yaml_generators re-exports them, so existing call
sites (main.rs, yq_bench.rs) are unchanged.

Prep for the next commit, which has benches/yq_comparison.rs include
this file directly via #[path].
